DTR-2311: test for request model
